+++ ghost/core/core/server/services/members/service.js
@@ -16,7 +16,7 @@ const models = require('../../models');
 const {GhostMailer} = require('../mail');
 const jobsService = require('../jobs');
 const tiersService = require('../tiers');
-const VerificationTrigger = require('@tryghost/verification-trigger');
+const VerificationTrigger = require('../VerificationTrigger');
 const DatabaseInfo = require('@tryghost/database-info');
 const settingsHelpers = require('../settings-helpers');
 const RequestIntegrityTokenProvider = require('./RequestIntegrityTokenProvider');
